Add SRI support for Node.js Runtime
Support for setting [Subresource
Integrity](https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity)
attributes on app router scripts was added in #39729. But this only
covered pages using the Edge Runtime.

With this PR, we're adding support for pages using the Node.js Runtime.
The only change that's needed for that, and which was probably just an
oversight in the original PR, is reading the generated manifest file in
`loadComponents`.